pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exa-py>=1.7.1",
    "lancedb>=0.17.0",
    "openai>=1.59.6",
    "phidata>=2.7.7",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "tantivy>=0.22.0",
]
